Log full HTML response to a tmpfile when debug is on
The full HTML response body is the single most valuable reproducer when
something goes wrong during the handling of that response.

This commit writes it to a tmpfile (usually under $TMPDIR, with the
googler-response- prefix) and prints the path to console when debug
(-d) is on:

    [DEBUG] Response body written to '/tmp/googler-response-oqY18e'.

so when asked a user could supply that log to facilitate debugging.