Set columns to 0 when sys.stderr is not a tty...
or otherwise fails fcntl.ioctl.

Currently, fcntl.ioctl fails with IOError (OSError on Python 3.x, where
IOError is an alias for OSError) when sys.stderr is not a terminal.
Although googler is mostly terminal oriented, we should not preclude
noninteractive use (it is an entirely conceivable use case to query
googler with a set of keywords and grep out all URLs in the output); and
even for interactive use, 2>/dev/null leading to an error isn't nice
UX either, especially considering that googler prints to stdout for the
most part. This failure also precludes any meaningful automated testing.

This commit removes the aforementioned limitation, and also fixes a
related comment (hard wrap != truncate, the latter implies loss of
information).