README.md: Clarify supported Python versions
Apparently there's no reason to support outdated patch releases, so
clarify that. The latest patch release of each minor version is what we
test on the CI anyway.

As for Python 3.0, 3.1 and 3.2, these are partially broken transitional
releases that are largely irrelevant now (might be less true for 3.2
since PyPy3 is still on CPython 3.2; but CPython is our only concern
here). googler might work with them (we're still testing 3.2 on the CI,
by the way), or not, we shouldn't care, factor into development or
provide support for them.

[ci skip]