Add script tests/ci-test-wrapper, and use it in .travis.yml
Implement the watchlist feature, where test is carried out only if
googler, tests, or a manually specified list of paths have changed.
This solves the problem of wasted tests when, say, only README is
changed; note that wasted tests lead us to Google suspicious activity
block from time to time, and that is really annoying.

(The other solution to the problem is "ci skip", but I hardly ever
remember to drop that tag on the first attempt.)