.travis.yml: Remove nightly from the Python matrix
Our code is simple enough that it's almost impossible to have
forward-compatibility problems. I'll test googler against upcoming
alphas, betas and release candidates anyway, so let's remove it from the
matrix to further reduce the number of queries we make, and thus
reducing the frequency of Google blocking us on Travis.