googler: quit_program: Use global conn instead of parameter
quit_program can only be called once, with the predestined parameter, so
it need not be a parameter. Note that other functions, e.g.,
fetch_results, also use the conn global.

We also add a kwarg exitcode to allow a custom exit code, which defaults
to 1 to preserve the original behavior.