googler: Add support for unfiltering filtered results
Sometimes Google omits similar (more like duplicate) result
entries. Whenever this happens, a notice is displayed in p#ofr that
reads like

    In order to show you the most relevant results, we have omitted some
    entries very similar to the 7 already displayed. If you like, you
    can repeat the search with the omitted results included.

The way to unfilter is simple. Just add '&filter=0' to the query string.

In this commit, we watch for p#ofr, and when detected, we set the
filtered property of the parser. Then after showing all results, we
display a notice and enable a contextual command 'unfilter' in the
prompt. Since it is rare and contextual, I think it is okay to give it a
long name.

Sample query: "why oh-my-zsh is completely broken" (quotes are part of
the query).