googler: Implement URL construction class GoogleUrl
The URL construction interface is now fully OO. It is to some extend
modeled on urllib.parse.ParseResult for familiarity, but it takes care
of all aspects of URL construction (even pagination) and defines a
easy-to-use and flexity interface with properties and methods. See the
docstring of the GoogleUrl class for details, as well as the docstrings
of individual user-facing methods.

As a side effect, we have reduced the number of globals, which are now
grouped into three categories: state carriers (which, as the name
suggests, carry global states), output/user interaction options, and
cosmetic options.