googler: Implement connection handling class GoogleConnection
GoogleConnection absorbed new_connection, google_get, and the connecting
and fetching parts of fetch_results. The new OO interface is nicely
abstract and fully documented.

Note that we have now registered an exit handler for closing the
connection. This way we can exit the program at any point using sys.exit
instead of our kind of awkward quit_program.

There's one breaking change which should not have much of a consequence:
we no longer modify the host segment of the global URL (now managed by
google_url) after a redirection. This is partly stipulated by the OO
interface (GoogleConnection class should not have access to the global
google_url) and partly because users should get what they specified when
they ask for the URL (with the 'o' command, for instance), not what
Google forced on them with a redirection.