Add devbin with the first dev script parse
devbin/ is supposed to be home to developer scripts.

The first script to be added is called `parse' which is intended for
experimenting with or testing GoogleParser alone, without moving
parts. It raises the usefulness of the saved responses from --debug to a
whole new level, because we can parse the saved response repeatedly
rather than just eyeballing the HTML.