Introduce "googler @t"
googler @t is a convenient add-on to search websites using small
unique keywords starting with the '@' symbol. This is a far more
convenient alternative to using the '-w' program option. To search
'hexspeak' on Wikipedia, run:

    @w hexspeak

Shell tab-completion after keying in the '@' makes it quite attractive
to use! Hence, adding the file under the `auto-completion` directory.