Custom URL opener other than a standard webbrowser
Use subprocess to spawn a custom script or CLI utility with the URL as param.
This is very useful when users want to do something different with the result
URLs.

For example:

Using a cleansing article reader delivers the best news reading experience.
To achieve that in the terminal I wrote a custom script:

```

elinks "http://justread.mpgarate.com/read?url=$1"
```

and named the utility `reader`. I am craving to use:

    googler -N sports --url-handler reader
to read the news!