Workaround #207, #218: support URL folding
We have explained in #207 our reservations regarding automatically adjusting the
number of search results according to the terminal size in googler. However, we
have the option to show the domain name only instead of the complete URL. This
is a visual change and many modern browsers do it. We cannot make it default at
this stage of the project. So I chose to use an environment variable. There's a
new omniprompt switch (`u`) to toggle it at runtime.