win32: also ENABLE_VIRTUAL_TERMINAL_PROCESSING on stderr
Follow-up on #277.

As discussed here[1], it's not really necessary when stdout and stderr are the
same console, but it doesn't hurt either.

Note that we did not test sys.stderr.isatty() in the code. This has been
considered, and deemed not a problem, since stdout is handled first (what
really matters) and we have a catch-all pass for any possible exception.

[1] https://github.com/jarun/googler/pull/277#discussion_r274245020