Disguise as IE
Currently when Google encounters an unknown user agent, it sometimes
responds with the mobile layout, which is hard to parse due to the use
of random-looking CSS classes.

We change the user agent to disguise as IE in order to retrieve the
classic layout we've been relying on.

Credit to mmbiz for reporting the issue (#284) and investigating various
user agent strings.

Closes #284.