Added the `--url-output` option.
Instead of running a new process for each selected URL
like with ``--url-handler``, `--url-output`` takes a file path,
which it writes the selected URLs to, delimited by '\n'.

This can be much more useful and efficient than url-handler.
url-handler requires an executable to run for each URL,
but does not allow any extra arguments to be sent to the program,
and since it must be an executable,
it can't be a bash function, for example.

url-output, when the file path is a named pipe,
allows for much more robust url handling.
The program that reads from the pipe as googler writes to it can do anything,
including running a bash function or passing extra args to it,
and it also avoids forking a new process for each URL.