Deduplicate results
Previously results may be duplicated, e.g. for the response
https://git.io/JJn05 the top result (from a featured snippet) is shown
in googler output twice.

The reason that happened is that a feature snippet could contain div.g
inside a div.g, so when we select results based on div.g we picked up
the same result twice -- the second container is a child of the first.

Instead of tracking node ancestry which is rather annoying, we introduce
__eq__ on Result and make sure no duplicate is recorded that way. Also
introduced __hash__, not actually in use but why not.