Replace uuid1 with uuid4 for privacy purpose
uuid1 is generated by mixing the host computers MAC address, and the
current time. The uniqueness comes at the cost of anonymity, which means
this may leak the machine's MAC address to Google.

Instead, uuid4 is generated randomly, and it almost impossible to occur the
conflict situation unless generating trillions of IDs every in a short time.

This patch was provided by Jakub Wilk [1]

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=965066

Signed-off-by: Jakub Wilk <jwilk@jwilk.net>
Signed-off-by: SZ Lin (林上智) <szlin@debian.org>