This fixes issue https://github.com/jarun/googler/issues/347 without using cookies.
Added the option to specifiy the geolocation used by google. This is simply
done by providing the "gl" parameter in the URL. The country codes are the
as the codes used in TLDs.

This might seem unimportant to people in english-speaking countries, but
searching computer related stuff is really annoying when it mixes results
from other languages.